split retry_task_map from otaclient_common.common
